requirements.txt
crewai
crewai[google-genai]
python-dotenv
fastapi
uvicorn
requests
pydantic
google-generativeai
loguru
sqlalchemy
psycopg2-binary